// src/app/model/skill.model.ts
export enum Skill {
    FRONT_END = "Dev front-end",
    BACK_END = "Dev back-end",
    MOBILE = "Dev mobile",
    DATA = "Data analysis / Dataviz",
    DEEP_LEARNING = "Deep learning",
    BLOCKCHAIN = "Blockahain",
    DEVOPS = "DevOps / Architecture",
    PRODUCT_OWNER = "Product owner",
    UX_UI = "UI/UX Design",
    ENTREPRENARIAT  = "Entrepreneuriat",
    ESS = "Institutions publiques",
    INVEST = "Investissement / Impact investing",
    DATA_SCIENCES = "Data science",
    OTHERS = "Autres",
    UNKNOW = "Je ne sais pas"
}